refactor: modularized edit videos on case and homologued styles
